+++ CHANGELOG.md
@@ -26,7 +26,9 @@ BUG FIXES:
 - Fix error in `getMessageWithCode()` in CustomError when error cause is undefined.
 - Clean up data in cache DB when create request and create identity fails.
 - Fix AS can send data response with any request ID and service ID without having to receive the request first.
+- Fix IdP can send response with any valid request ID without having to receive the request first.
 - Fix process exits on MQ error by handling error events emitted from MQSend and MQRecv.
+- Fix miscommunicated error response when RP trying to create a request with AS ID that does not provide the requested service.
 
 ## 0.5.4 (July 27, 2018)